Update Code.py
Now-a-days there are a lot of works in a company so the head of the company thought why not
to automate the information through mail. He wants a system through which a mail informing
the employees about holiday and weekends, also some of special holidays given by the company.
The mail should be sent to all persons, the email ids with their names, and other information
will be present in the excel file. The new changes have been done by Kalpak Kanungo.